pyproject.toml
[project]
name = "databrick-iot"
version = "0.1.0"
description = "IoT Data Generator and Analytics Pipeline for Databricks"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-storage-file-datalake>=12.8.0",
    "azure-identity>=1.12.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.0",
    "mlflow>=3.4.0",
    "seaborn>=0.13.2",
    "plotly>=6.3.1",
]

[dependency-groups]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing"
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0"
]
